pixel spacing fixed, can save realm dat and copy to app bundle
