+++ BlueStarRedStar/AppDelegate.swift
@@ -1,5 +1,6 @@
 
 import UIKit
+import RealmSwift
 
 @UIApplicationMain
 class AppDelegate: UIResponder, UIApplicationDelegate {
@@ -9,6 +10,11 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
 
   func application(_ application: UIApplication, didFinishLaunchingWithOptions launchOptions: [UIApplication.LaunchOptionsKey: Any]?) -> Bool {
     // Override point for customization after application launch.
+    
+    var config = Realm.Configuration()
+          config.fileURL =  Bundle.main.url(forResource: "default", withExtension: "realm")
+          
+          Realm.Configuration.defaultConfiguration = config
     return true
   }
