+++ Shared Files/Realm Objects/Segment.swift
