+++ TwinStarLevelEditor/Base.lproj/Main.storyboard
@@ -60,6 +60,12 @@
                                 <modifierMask key="keyEquivalentModifierMask"/>
                                 <menu key="submenu" title="File" id="bib-Uj-vzu">
                                     <items>
+                                        <menuItem title="Copy to app" id="Ukd-fX-3QN">
+                                            <modifierMask key="keyEquivalentModifierMask"/>
+                                            <connections>
+                                                <action selector="copyToApp:" target="Ady-hI-5gd" id="q0U-e0-jm3"/>
+                                            </connections>
+                                        </menuItem>
                                         <menuItem title="Save" id="i3f-zi-CYm">
                                             <modifierMask key="keyEquivalentModifierMask"/>
                                             <connections>
